+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - Fixed an issue where cells were not recalculated after adding new sheet. [#1116](https://github.com/handsontable/hyperformula/issues/1116)
+- Fixed an issue where overwriting a non-computed cell caused the `Value of the formula cell is not computed` error. [#1194](https://github.com/handsontable/hyperformula/issues/1194)
 
 ## [3.1.0] - 2025-10-14